ApplyPropagation

This function applies an in-silico perturbation to selected features in a Seurat object.

Usage

ApplyPropagation(
  seurat_obj,
  exp,
  exp_per,
  network,
  perturb_dir = perturb_dir,
  n_iters = 3,
  delta_scale = 0.2
)

Arguments

seurat_obj

A Seurat object

exp

A features by cells matrix containing the observed expression matrix.

exp_per

A features by cells matrix containing the perturbation results from ApplyPerturbation.

network

A gene-gene network to apply the signal propagation.

n_iters

The number of times to apply the signal propagation.

Value

A dgCMatrix object containing the updated expression matrix with the applied perturbations